Move config operation into separate file
